Progress #1281 -- Minimizing TeamHasVisionOn and network usage (#1296)
The TeamHasVisionOn function itself is fast, but was called in three places per frame for the same objects:

Screenshot 2022-02-04 at 23-59-25 ms_circle_quadtree speedscope - speedscope - annotated

Therefore, all usages have been replaced with GameObject.IsVisibleByTeam. But since the BroadcastPacketVision function used when spawning objects requires up-to-date information about visibility, it was decided to restructure ObjectManager.Update so that first the obj.Update takes place, then the visibility calculation for all objects created during update and old objects, then (at the end of the frame) new objects are spawned and old ones are synchronized.

Apparently, due to the fact that earlier objects spawned only at the beginning of the next frame, the Particles had the autoSend option. I decided not to make exceptions for them, and just comment it out.

I also commented out the extra NotifyUpdatedStats calls marked with a // TODO: send this in one place only comment, because everything is already being done this way. And in the Update function, only changed stats are now synchronized.

And in UpdateVision, I fixed a bug that I made, which led to the fact that the visibility status of always visible objects, such as towers, was constantly updated 😅

* collision optimizations for old QuadTree

* Dissolve Particle VisionAffected into Vision Checks

* Fix Missed Rebase Changes

* NEUTRAL Regions Give Global Vision

* Fix Global Vision

Co-authored-by: microsoftv <6063922+microsoftv@users.noreply.github.com>