Progress #1281 -- The return of meaning to the `_visionUnits` (#1302)
During the optimizations, `_visionUnits` and all the infrastructure associated with them were left unused, probably due to the fact that they were not used as originally intended from the very beginning. Instead, the concept of `_visionProviders` has been introduced - units that can, under certain circumstances, provide a view of the unit to their team. Now they are used when determining whether a command can see a unit, which reduces the cost of iterating over all `_objects`. The right to provide vision is assigned to `IAttackableUnit`s and `Region`s, with the proviso that `TEAM_NEUTRAL` regions provide vision to both teams

* TP command has been fixed to not crash on incorrect NetID.
* Added support for regions in ObjectManager.TeamHasVisionOn

Co-authored-by: microsoftv <6063922+microsoftv@users.noreply.github.com>