Progress #1281 -- Another QuadTree implementation (#1303)
I have modified and integrated another implementation of `QuadTree`. It is faster and sharpened for working with `Circle`s, and most importantly, it allows you to do what a tree should allow - get a list of objects in a given radius, which will eventually replace functions like `GetUnitsInRange` that iterate over all `_objects`.