Quickfix: non-removable missiles (#1377)
Missiles are only removed when they collide with their target. If the target dies early, the missile stays and continues to check its visibility. Ordinary missiles don't have to check for collision at all, but simply move towards the enemy or the last point where he was, and then disappear. And perhaps should not check the vision. But this is a topic for future commits.