Progress #1281 -- Removing unnecessary conversions and double cell search from CastRay (#1378)
For some reason, `CastRay` did not use the coordinates of the grid cells, but the real coordinates, which are about 5 times larger, which is why the function made more cycles. Also, thanks to the translation of the function to the use of converted coordinates, it became possible not to do this once again. Plus, when checking visibility, the cell is now searched only once. I also switched to using vectors in the hope that someday the compiler will be able to optimize working with them.