Progress #1366 -- Various TeleportID-related fixes (#1367)
I wrote "progress" because I still have to use teleportation for collisions. Otherwise, collision avoidance comes into play and the minions go in different directions. They are already trying to do this, but in the *indev* branch they are teleported away every tick of the server. This also looks ugly, but I decided to keep this behavior.
At least now, instead of recalculating the entire path, a collision only changes the zero point of the path. And this is usually enough, because the collision only throws the unit back a little.