`PlayerID` vs `ClientID` (#1468)
In its packets, the League uses concepts such as:
`long PlayerID` - identifies the player's account
`int ClientID` (from 0 to n - 1, where n is the number of players) - identifies the player in the current game.
In the Sandbox, instead of `ClientID` (often referred to as `userId`), `PlayerID` is used, which causes confusion and uncontrollable casting from `long` to `int` (and also `ulong` and `uint`) and vice versa, which can create problems in the long run.
On this occasion, TODOs are found in the code, calling to stop this obscurantism.
In addition, things like `List<Tuple<uint, ClientInfo>>` are used. The `ClientInfo` structure already contains the `ClientId`, and the number of players in some places is known in advance - no need to be so perverse.
I was inspired by the `LeagueServer` class. Now the correct terminology and ranges are used everywhere.